test: Add mock fixtures for MongoDB and product repository tests
